simplify: extract shared stripInternalColumns, use getSourceCheckHref, dedupe sets
- Extract stripInternalColumns to shared module (was duplicated in
  entity-profile.ts and record-lookup.ts)
- Use getSourceCheckHref() helper instead of inline URL construction
  in things detail page (2 sites)
- Derive VERDICT_SOURCE_TABLES from SOURCE_TABLE_TO_RECORD_TYPE keys
  instead of maintaining parallel hand-written set
- Remove unused ArrowLeft import (already done in prior commit — no-op)
