+++ apps/wiki-server/src/routes/tablebase/entity-profile.ts
@@ -33,6 +33,7 @@ import { isAnySid } from "@longterm-wiki/id-utils";
 import { resolveEntityStableId } from "../shared/entity-resolution.js";
 import { notFoundError } from "../shared/utils.js";
 import { COLUMN_DESCRIPTIONS } from "./entity-profile-descriptions.js";
+import { stripInternalColumns } from "../shared/strip-internal-columns.js";
 import type { PgTable } from "drizzle-orm/pg-core";
 
 const logger = rootLogger.child({ component: "entity-profile" });
@@ -69,7 +70,7 @@ function buildSchemaForTable(table: PgTable, tableName: string): ColumnMeta[] {
   });
 }
 
-// ---- Columns to exclude from output (internal/noisy) ----
+// ---- Columns to exclude from schema output (still needed for buildSchemaForTable filter) ----
 
 const EXCLUDED_COLUMNS = new Set([
   "synced_at",
@@ -80,23 +81,6 @@ const EXCLUDED_COLUMNS = new Set([
   "synced_from_branch",
   "synced_from_commit",
 ]);
-
-// Also build a camelCase version for stripping from Drizzle query results
-const EXCLUDED_CAMEL = new Set(
-  [...EXCLUDED_COLUMNS].map((k) => k.replace(/_([a-z])/g, (_, c) => c.toUpperCase()))
-);
-
-function stripInternalColumns(rows: Record<string, unknown>[]): Record<string, unknown>[] {
-  return rows.map((row) => {
-    const cleaned: Record<string, unknown> = {};
-    for (const [key, value] of Object.entries(row)) {
-      if (!EXCLUDED_CAMEL.has(key)) {
-        cleaned[key] = value;
-      }
-    }
-    return cleaned;
-  });
-}
 
 // ---- Query safety ----
